+++ test/engine_test.rb
@@ -16,6 +16,6 @@ class EngineTest < Minitest::Test
 
   def test_recording_studio_dependency_is_loaded
     assert defined?(RecordingStudio)
-    assert_operator RecordingStudio::VERSION, :>=, "3.0.0"
+    assert_operator RecordingStudio::VERSION, :>=, "4.2.0"
   end
 end